+++ mssp-20161228/ChangeLog.md
@@ -1,3 +1,12 @@
+2024-12-19 Version: 1.1.0
+- Support API ConfirmDjbhReport.
+- Support API CreateServiceLinkedRole.
+- Support API DeleteDjbhReport.
+- Support API DescribeServiceLinkedRole.
+- Update API GetDocumentDownloadUrl: add param ReportType.
+- Update API GetDocumentPage: update response param.
+
+
 2024-12-10 Version: 1.0.0
 - Support API GetConsoleScore.